+++ src/focusable.js
@@ -39,6 +39,8 @@ export function Focusable(
     onSelect,
     onBack,
     onMove,
+    onFocus,
+    onBlur,
 
     ...rest
   },
@@ -68,7 +70,7 @@ export function Focusable(
     };
   });
 
-  const { nodes, createNode, destroyNode } = parentProviderValue;
+  const { nodes, createNode, destroyNode, updateNode } = parentProviderValue;
 
   const possibleNode = nodes[idRef.current];
   const hasNode = Boolean(possibleNode);
@@ -115,6 +117,8 @@ export function Focusable(
         onDown,
         onSelect,
         onBack,
+        onFocus,
+        onBlur,
 
         onMove,
       });
@@ -140,6 +144,15 @@ export function Focusable(
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
+
+  useEffect(() => {
+    if (hasNodeRef.current) {
+      updateNode(idRef.current, {
+        disabled,
+      });
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [disabled]);
 
   const classString = `${className} ${isFocused ? focusedClass : ''} ${
     isFocusedExact ? focusedExactClass : ''
@@ -194,6 +207,8 @@ ForwardedFocusable.propTypes = {
   onSelect: PropTypes.func,
   onBack: PropTypes.func,
   onMove: PropTypes.func,
+  onFocus: PropTypes.func,
+  onBlur: PropTypes.func,
 };
 
 export default ForwardedFocusable;